+++ vset/memory/fsb/aspect/ad_pred_offset.hpp
@@ -31,7 +31,7 @@ struct ad_pred_offset
     {
       if ( value_type* current = chn->last_value() )
       {
-        offset = reinterpret_cast<data_type>(current) - data;
+        offset = static_cast<size_t>( reinterpret_cast<data_type>(current) - data );
       }
     }
     else for (size_t i = 0 ; i < count ; ++i)
@@ -39,7 +39,7 @@ struct ad_pred_offset
       
       if ( value_type* current = chn->pred_value( reinterpret_cast<value_type*>(data + offset) ) )
       {
-        offset = reinterpret_cast<data_type>(current) - data;
+        offset = static_cast<size_t>( reinterpret_cast<data_type>(current) - data );
       }
       else
       {
